fixed some build warnings
git-svn-id: http://crashrpt.googlecode.com/svn/trunk@1383 821a4644-a21a-0410-aec7-6199d47568cc

+++ reporting/crashsender/MailMsg.cpp
@@ -261,7 +261,7 @@ BOOL CMailMsg::Send()
     message.lpszConversationID                = NULL;
     message.flFlags                           = 0;
     message.lpOriginator                      = pRecipients;
-	message.nRecipCount                       = m_to.size();
+	message.nRecipCount                       = (ULONG)m_to.size();
     message.lpRecips                          = &pRecipients[1];
     message.nFileCount                        = nAttachments;
     message.lpFiles                           = nAttachments ? pAttachments : NULL;
